python/requirements.txt
opencv-python
Pillow
imagehash
reportlab
flask
flask-cors
numpy
watchdog